requirements.txt
streamlit
pandas
numpy
pillow
pymupdf        # for fitz
easyocr
ultralytics
torch
torchvision
opencv-python-headless
xlsxwriter
streamlit
pandas
numpy
Pillow
PyMuPDF
easyocr
ultralytics
torch
torchvision
opencv-python-headless
xlsxwriter
